requirements.txt
django
django-widget-tweaks
Django==3.1.3
djangorestframework==3.12.2
libgravatar==0.2.4
pytz==2020.4
django-filter==2.4.0
